Add new component: IndividualArticle.
