+++ src/Components/AllArticles/AllArticles.js
@@ -1,23 +1,11 @@
-const AllArticles = ({ articles }) => {
-  const allArticles = articles.map((article) => {
-    return (
-      <div className="article-container">
-        <div className="title-author">
-          <h1>
-            {article.title}
-          </h1>
-          {article.author}
-        </div>
-        <img src={article.urlToImage} alt={article.title} className="article-img"></img>
-      </div>
-    )
-  })
-  
+import IndividualArticle from "../IndividualArticle/IndividualArticle";
+
+const AllArticles = ({ articles }) => {  
   return (
-    <div>
-      {allArticles}
+    <div className="all-articles">
+      <IndividualArticle articles={articles}/>
     </div>
   )
-}
+};
 
 export default AllArticles;
