+++ src/Components/App/App.css
@@ -1,27 +1,29 @@
-.App {
-  text-align: center;
-}
-
 .navbar {
   width: 100%;
-  height: 30vh;
+  height: 60vh;
   background-image: url('../../images/newspaper-printing.png');
   background-size: 100% auto;
+  color: white;
+  font-size: 80px;
+  text-align: left;
 }
 
 .article-container {
-  background-color: blanchedalmond;
-  padding: 10px;
-  margin: 10px;
+  padding: 60px;
+  margin: 40px;
   display: flex;
-  /* flex-direction: row; */
 }
 
 .article-img {
-  width: 25vw;
+  width: 35vw;
 }
 
 .title-author {
-  width: 75vw;
-  padding: 50px;
+  width: 35vw;
+  padding: 30px;
+  text-align: right;
+}
+
+h1 {
+ margin: 0;
 }
